refactor(crashreporting): rename crash_reporting to crashreporting
